requirements.txt
haystack-ai
faiss-cpu
pypdf
openai
tqdm
